added sendable properties to subsystems, fixed constants, cleaned up function names

+++ src/main/java/com/stuypulse/robot/commands/DrivetrainAlignmentCommand.java
@@ -3,7 +3,6 @@ package com.stuypulse.robot.commands;
 import com.stuypulse.robot.Constants.Alignment;
 import com.stuypulse.robot.subsystems.Drivetrain;
 import com.stuypulse.stuylib.control.Controller;
-import com.stuypulse.stuylib.control.PIDController;
 import com.stuypulse.stuylib.math.Angle;
 import com.stuypulse.stuylib.math.SLMath;
 import com.stuypulse.stuylib.network.limelight.Limelight;
@@ -27,13 +26,18 @@ public class DrivetrainAlignmentCommand extends DrivetrainCommand {
     public interface Aligner {
         // Called when command initialize is called,
         // Useful for relative encoder commands
-        public default void init() { }
+        public default void init() {
+        }
 
         // The amount of positional error
-        public default double getSpeedError() { return 0.0; };
+        public default double getSpeedError() {
+            return 0.0;
+        };
 
         // The amount of angular error
-        public default Angle getAngleError() { return Angle.fromDegrees(0); };
+        public default Angle getAngleError() {
+            return Angle.fromDegrees(0);
+        };
     }
 
     // Max speed for the robot
@@ -114,7 +118,7 @@ public class DrivetrainAlignmentCommand extends DrivetrainCommand {
         this.maxSpeed = speed;
         return this;
     }
-    
+
     // Make the command never finish
     public DrivetrainAlignmentCommand setNeverFinish() {
         this.neverFinish = true;
@@ -144,16 +148,20 @@ public class DrivetrainAlignmentCommand extends DrivetrainCommand {
         aligner.init();
         timer.reset();
 
-        this.speed.setErrorFilter(new LowPassFilter(Alignment.Speed.IN_SMOOTH_FILTER.doubleValue()));
-        this.speed.setOutputFilter(new IFilterGroup(
-            (x) -> SLMath.limit(x, maxSpeed),
-            new LowPassFilter(Alignment.Speed.OUT_SMOOTH_FILTER.doubleValue())
-        ));
+        this.speed.setErrorFilter(new LowPassFilter(Alignment.Speed.IN_SMOOTH_FILTER));
+        this.speed.setOutputFilter(
+            new IFilterGroup(
+                (x) -> SLMath.limit(x, maxSpeed),
+                new LowPassFilter(Alignment.Speed.OUT_SMOOTH_FILTER)
+            )
+        );
 
-        this.angle.setErrorFilter(new LowPassFilter(Alignment.Angle.IN_SMOOTH_FILTER.doubleValue()));
-        this.angle.setOutputFilter(new IFilterGroup(
-            new LowPassFilter(Alignment.Angle.OUT_SMOOTH_FILTER.doubleValue())
-        ));
+        this.angle.setErrorFilter(new LowPassFilter(Alignment.Angle.IN_SMOOTH_FILTER));
+        this.angle.setOutputFilter(
+            new IFilterGroup(
+                new LowPassFilter(Alignment.Angle.OUT_SMOOTH_FILTER)
+            )
+        );
 
         updateTargets();
         pollingTimer.reset();
@@ -167,7 +175,7 @@ public class DrivetrainAlignmentCommand extends DrivetrainCommand {
 
     // Get distance left to travel
     public double getSpeedError() {
-        if(this.useInterpolation) {
+        if (this.useInterpolation) {
             return targetDistance - drivetrain.getDistance();
         } else {
             return aligner.getSpeedError();
@@ -176,7 +184,7 @@ public class DrivetrainAlignmentCommand extends DrivetrainCommand {
 
     // Get angle left to turn
     public Angle getAngleError() {
-        if(this.useInterpolation) {
+        if (this.useInterpolation) {
             return targetAngle.sub(drivetrain.getAngle());
         } else {
             return aligner.getAngleError();
@@ -211,7 +219,7 @@ public class DrivetrainAlignmentCommand extends DrivetrainCommand {
         super.execute();
 
         // Update targets if time has come
-        if(pollingTimer.getTime() > Alignment.INTERPOLATION_PERIOD) {
+        if (pollingTimer.getTime() > Alignment.INTERPOLATION_PERIOD) {
             updateTargets();
             pollingTimer.reset();
         }
@@ -220,22 +228,22 @@ public class DrivetrainAlignmentCommand extends DrivetrainCommand {
     // Command is finished if all of the errors are small enough
     public boolean isFinished() {
         // If you do not want the command to automatically finish
-        if(neverFinish) {
+        if (neverFinish) {
             return false;
         }
 
         // Check if the aligner hasn't run for long enough
-        if(this.minTime && timer.getTime() < Alignment.MIN_ALIGNMENT_TIME) {
+        if (this.minTime && timer.getTime() < Alignment.MIN_ALIGNMENT_TIME) {
             return false;
         }
 
         // If continuous, do not check for velocity
-        if(this.continuous) {
-            return (speed.isDone(Alignment.Speed.MAX_SPEED_ERROR * 2.5) 
-                 && angle.isDone(Alignment.Angle.MAX_ANGLE_ERROR * 1.5));
+        if (this.continuous) {
+            return (speed.isDone(Alignment.Speed.MAX_SPEED_ERROR * 2.5)
+                    && angle.isDone(Alignment.Angle.MAX_ANGLE_ERROR * 1.5));
         } else {
-            return (speed.isDone(Alignment.Speed.MAX_SPEED_ERROR, Alignment.Speed.MAX_SPEED_VEL) 
-                 && angle.isDone(Alignment.Angle.MAX_ANGLE_ERROR, Alignment.Angle.MAX_ANGLE_VEL));
+            return (speed.isDone(Alignment.Speed.MAX_SPEED_ERROR, Alignment.Speed.MAX_SPEED_VEL)
+                    && angle.isDone(Alignment.Angle.MAX_ANGLE_ERROR, Alignment.Angle.MAX_ANGLE_VEL));
         }
     }
